requirements.txt
torch
numpy
gym
pynput
matplotlib
stable_baselines3[extra]
wandb
hydra-core